+++ src/app/edit-roles/components/index.tsx
@@ -1,3 +1,3 @@
 import RoleName from './name/rolename';
-import EditTable from './table/edittable';
+import EditTable from './table/roletable';
 export  { RoleName, EditTable };
